视频模块基本完成,添加、删除、修改、上传视频 帧图获取还未完成，限制视频上传大小最大为20MB，头像大小最大为2MB，增加视频删除判断，视频在线预览判断

+++ src/main/java/cn/annna/service/UserServiceImpl.java
@@ -121,6 +121,9 @@ public class UserServiceImpl implements UserService{
             if (!Objects.equals(headImg.getContentType(), "image/jpeg") && !Objects.equals(headImg.getContentType(), "image/png")){
                 throw new RuntimeException("只允许上传 jpeg|png 格式的头像,请规范操作");
             }
+            if (headImg.getSize() > 2097152){
+                throw new RuntimeException("头像大小大于 2MB ,请调整");
+            }
             if (!oldHeadImg.equals("")){
                 OSSUtil.deleteFile(oldHeadImg);
                 String headPath = OSSUtil.uploadFile(headImg,"yingxue/images/head/");
